requirements.txt
Flask
supabase
python-dotenv
gunicorn
httpx